baseApp-main/requirements.txt
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dnspython==2.3.0
Flask==2.2.2
fonttools==4.37.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.0
Pillow==9.2.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1

